+++ lib/compiler/src/sema/passes/type_checker.cc
@@ -461,6 +461,7 @@ auto type_checker::check_instruction(gir::function& fn, const gir::instruction& 
                             const auto arg_t{get_operand_type(inst.operands[i])};
                             if (arg_t && !arg_t->is_poison() &&
                                 arg_t->get_kind() != sema::type_kind::TYPE &&
+                                params[i]->type.get_kind() != sema::type_kind::TYPE &&
                                 !is_assignable(*arg_t, params[i]->type)) {
                                 emit_diagnostic(
                                     fmt::format("Argument {} of type '{}' is not assignable to "
@@ -487,6 +488,7 @@ auto type_checker::check_instruction(gir::function& fn, const gir::instruction& 
                         const auto arg_t{get_operand_type(inst.operands[i])};
                         if (arg_t && !arg_t->is_poison() &&
                             arg_t->get_kind() != sema::type_kind::TYPE &&
+                            params[i]->type.get_kind() != sema::type_kind::TYPE &&
                             !is_assignable(*arg_t, params[i]->type)) {
                             emit_diagnostic(
                                 fmt::format("Argument {} of type '{}' is not assignable to "